backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.7.1
pydantic==2.4.2
pydantic-core==2.10.1
python-multipart==0.0.6